backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
python-multipart==0.0.6
pyserial==3.5
mpremote==1.26.1